+++ src/main/java/com/programmers/jpa/customer/dto/CreateRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.jpa.customer.ui;
+package com.programmers.jpa.customer.dto;
 
 public record CreateRequest(String firstName, String lastName) {
 }
