+++ src/main/java/com/programmers/jpa/customer/dto/UpdateRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.jpa.customer.ui;
+package com.programmers.jpa.customer.dto;
 
 public record UpdateRequest(Long id, String firstName, String lastName) {
 }
