+++ src/main/java/com/programmers/jpa/customer/ui/CustomerController.java
@@ -2,6 +2,9 @@ package com.programmers.jpa.customer.ui;
 
 
 import com.programmers.jpa.customer.application.CustomerService;
+import com.programmers.jpa.customer.dto.CreateRequest;
+import com.programmers.jpa.customer.dto.FindResponse;
+import com.programmers.jpa.customer.dto.UpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
